requirements.txt
thrift
gevent
